+++ hydra-cardano-api/src/Hydra/Cardano/Api.hs
@@ -113,7 +113,6 @@ import Hydra.Cardano.Api.Lovelace as Extras
 import Hydra.Cardano.Api.MultiAssetSupportedInEra as Extras
 import Hydra.Cardano.Api.NetworkId ()
 import Hydra.Cardano.Api.PlutusScript as Extras
-import Hydra.Cardano.Api.PlutusScriptVersion as Extras
 import Hydra.Cardano.Api.PolicyId as Extras
 import Hydra.Cardano.Api.ReferenceScript as Extras
 import Hydra.Cardano.Api.ScriptData as Extras